# MONGOID-5617 - Increase performance of `Mongoid::Matcher#extract_attribute` by avoiding BSON::Document instantiation