https://jira.mongodb.org/browse/MONGOID-5617

`Mongoid::Matcher#extract_attribute` creates a `BSON::Document` for indifferent access of hash keys.

Performance can be improved by explicitly attempting string and symbol access.

```console
$ ruby -v
ruby 2.7.3p183 (2021-04-05 revision 6847ee089d) [arm64-darwin21]

$ bundle exec ruby perf/benchmark_matches.rb 2>/dev/null
Warming up --------------------------------------
_matches? with BSON::Document
                         9.424k i/100ms
_matches? with indifferent key methods
                         9.999k i/100ms
Calculating -------------------------------------
_matches? with BSON::Document
                         94.447k (± 0.8%) i/s -    480.624k in   5.089113s
_matches? with indifferent key methods
                         99.887k (± 0.7%) i/s -    499.950k in   5.005367s

Comparison:
_matches? with indifferent key methods:    99887.4 i/s
_matches? with BSON::Document:    94447.0 i/s - 1.06x  slower
```